jellyfish-cli/com.ngc.seaside.jellyfish.api/src/test/java/com/ngc/seaside/jellyfish/api/CommonParametersTest.java
/**
 * UNCLASSIFIED
 *
 * Copyright 2020 Northrop Grumman Systems Corporation
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of
 * this software and associated documentation files (the "Software"), to deal in
 * the Software without restriction, including without limitation the rights to use,
 * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
 * Software, and to permit persons to whom the Software is furnished to do so,
 * subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
 * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
 * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
package com.ngc.seaside.jellyfish.api;

import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class CommonParametersTest {

   @Test
   public void testDoesParseGav() throws Throwable {
      String[] parsed = CommonParameters.parseGav("group:artifact:1.0");
      assertEquals("group ID not correct!",
                   "group",
                   parsed[0]);
      assertEquals("artifact ID not correct!",
                   "artifact",
                   parsed[1]);
      assertEquals("version not correct!",
                   "1.0",
                   parsed[2]);
   }

}
